criando fluxo para cadastrar nivel e foodres

+++ src/main/resources/db/migration/V3__create_table_level_folders.sql
@@ -3,6 +3,6 @@ CREATE TABLE level_folders
     id               UUID PRIMARY KEY,
     level_profile_id UUID         NOT NULL REFERENCES level_profiles (id) ON DELETE CASCADE,
     name             VARCHAR(100) NOT NULL,
-    position         SMALLINT     NOT NULL DEFAULT 0,
-    initial_files    SMALLINT     NOT NULL DEFAULT 0
+    position         INT          NOT NULL DEFAULT 0,
+    initial_files    INT          NOT NULL DEFAULT 0
 );
