+++ src/main/resources/db/migration/V5__create_table_folders.sql
@@ -3,6 +3,6 @@ CREATE TABLE folders
     id             UUID PRIMARY KEY,
     student_id     UUID          NOT NULL REFERENCES students (id),
     name           VARCHAR(100)  NOT NULL,
-    position       SMALLINT      NOT NULL DEFAULT 0,
+    position       INT      NOT NULL DEFAULT 0,
     created_at     TIMESTAMPTZ   NOT NULL DEFAULT NOW()
 );
